change ExternalDataSource into a StrEnum
