+++ app/db/types.py
@@ -56,7 +56,6 @@ class EntityType(StrEnum):
     experimental_bouton_density = auto()
     experimental_neuron_density = auto()
     experimental_synapses_per_connection = auto()
-    external_data_source = auto()
     external_data_source_page = auto()
     ion_channel_model = auto()
     memodel = auto()
@@ -494,3 +493,9 @@ class CircuitScale(StrEnum):
     region = auto()
     system = auto()
     whole_brain = auto()
+
+
+class ExternalDataSource(StrEnum):
+    channelpedia = "Channelpedia"
+    modeldb = "ModelDB"
+    icgenealogy = "ICGenealogy"
