+++ app/routers/__init__.py
@@ -20,7 +20,6 @@ from app.routers import (
     experimental_bouton_density,
     experimental_neuron_density,
     experimental_synapses_per_connection,
-    external_data_source,
     external_data_source_page,
     ion_channel_model,
     license,
@@ -68,7 +67,6 @@ authenticated_routers = [
     experimental_bouton_density.router,
     experimental_neuron_density.router,
     experimental_synapses_per_connection.router,
-    external_data_source.router,
     external_data_source_page.router,
     ion_channel_model.router,
     license.router,
